Add jparse_ prefix to json parser
This is so that one can have multiple parsers in a program that links in
the jparse library. This took extra effort for some reason but it should
be okay. Nevertheless we have yet to test jparse in other applications.
New version of jparse is 1.0.1 2023-03-10.

Add jparse.c to make depend list in jparse/Makefile and run make depend.
The file jparse.c was missing in the variable BUILT_C_SRC and this was
potentially causing problems as one of the issues in working out the
jparse_ prefix (which really shouldn't have been as complicated as it
turned out to be).

Updated CHANGES.md for these changes as well as the fix in
jparse/Makefile in commit 18cd69d63aa57b82ce6dfd78094b68f40ad21e0c made
yesterday.