Fix writing of answers file (past winner, handle)
When the past winner and winner handle prompts were added to the
mkiocccentry tool the tool was changed to read the values from the
answers file but it did not write it to the answers file.

But why did the test script work? It's simply because the script (of
course) does not run the interactive tool but merely reads from the
answers file which the script itself creates.

Now one can use the -a/-A feature and expect -i to work.