Fix warnings compiling jparse
Two warnings were being triggered under macOS: one about unused but set
variable (in generated code). That was solved by adding
-Wno-unused-but-set-variable to the WARN_FLAGS. Note that clang in some
platforms (like CentOS 7) do not have that option but it is not fatal so
this is okay (though in that platform it would generate additional
warnings).

The other was a more recent one:

    json_util.c:129:18: warning: a function declaration without a prototype is deprecated in all versions of C [-Wstrict-prototypes]
    json_warn_allowed()
		     ^
		      void

The solution is to add void to the function arg list.

Aesthetic changes that don't really matter but were done and by accident
made in this commit (I already saved and quit - this is an emendation):
off-by-one char in a comment and better wording of a comment as well.