Add remarks.example.md for authors
This file gives instructions for writing a remarks.md file for IOCCC
entries. This commit was supposed to be done quite a while ago but it
slipped my mind and I haven't looked at the repo in a long while (or git
status of the repo in a long while).

This file is of the same nature as the Makefile.example.

Updated CHANGES.md to note this addition.