Initial jprint code
Right now it only links in the jparse, dynamic array and debug
libraries and when run it exits 0.

The Makefile has been updated and make depend was run as well.

This tool will I believe take some time and careful thought and it
definitely will be developed over many commits but now it should be
possible to just start writing the code without having to worry about
editing the Makefile etc. More editing of the Makefile will have to be
done later - for instance when man pages are added or some dependency
changes or if a new file has to be added but at this point I can start
working on the code once I have a better idea of what is needed. This
won't happen until a discussion is had on GitHub though I might be able
to get some code in prior to that should I find the time and energy. I
do hope to do a bit more on this later today.

Updated .gitignore: add jparse/jprint.