Add initial INCOMPLETE command line parsing
Nothing specific to this tool is parsed yet and nothing in that
category can be until more discussion is had.

This is just to get this part out of the way as I started it just to get
the initial part done. Before we can continue working on the actual code
of this tool we have to discuss the tool more including the usage
message. These options have the standard (for this repo) options but
nothing more.

Without -h no usage message is printed but this is because nothing is
done beyond parsing args. Nothing can be done until more is discussed in
detail.

It's possible that no further updates to these files will be done for
some days at the very least for the above reasons.