Update jprint usage string + check required args
The usage string is incomplete per the comment in GitHub but this is
because it's both long (and so the question of whether the details for
the remaining options unexplained should have more than just a single
line is to be considered, see below) and I am currently too tired to
finish putting in the explanation of the options and the default values
anyway: this can come later.

However some of the basic options are explained and the synopsis is
there. In a future commit a man page will need to be made as well but
that's for another commit in another pull request.

Required number of args is 2.

Updated jprint version.