New jprint version - 'complete' help string
After many, many, many (many, many, many!) :-) new inodes[0], the usage
string is at this time complete: for certain values of complete :-)
There are some things that are unclear to me but these will be ironed
out with discussion on GitHub (not that it's physically on or IN GitHub
:-) ) and then the usage string can be fixed.

Why a new version? Because it is significant even if there is no
functional addition (usage string was always printed).

Note that not even the new options are accepted and none of the options
accepted are actually parsed. This can come later once details are
ironed out though I possibly will add them to the getopt() call and
switch() cases - but not right now.

[0] To those who do not know every time you save a file it creates a new
inode for the same file. If you don't understand why this is I suggest
you look it up (as I'm not about to explain since I've been awake since
stupid o'clock and it's still stupid o'clock and I want to try and rest
soon).