Fix rest of location.1
I believe that at this time everything is good. There might be a few
things off / wrong but there were additional typo fixes and a minor
change in an example.

I am actually aware that the EXAMPLES section could probably be improved
formatting wise but since it looks fine it's probably okay and it might
be that I am remembering wrong: certainly it's the format I once had but
I believe I ended up changing it a bit later on. This is unneeded
however.