Initial parsing of options in jprint
This commit is an important step in progressing on the tool as now we
can discuss what needs to be done, what has to change etc. wrt the
options and functionality of the tool. What this commit includes is as
follows.

The getopt() call should be complete but there are few comments except
in the variables for the options. Note that some variable types will
possibly have to change and as far as option arguments being parsed only
the integers and booleans are complete. As the rest of the details
become clearer more can be done.

A word on the usage string. Right now I have the longer version as it
will be useful to have the information when I work more on the code.
As each part is done the usage string in that part can be simplified

I also added the exit codes to the usage string as requested on GitHub.