Typo fix jprint.c comment
I also want to say a couple things that were not mentioned in the
previous commit that were added or changed.

Most importantly one can now specify more than one name arg.

Also the -c option to jprint was changed to -T and -c was given the same
meaning as grep -c.