New jprint version - allow setting max tree depth
Specifying -m depth sets the maximum depth the tool will traverse in the
JSON tree. Used only if the json is valid.

Explicitly include dyn_array.h and update Makefile to be clear on this.

Updated CHANGES.md for new jprint version. Fixed date of previous
release as well: it said it was 5 June but that's today, not yesterday.