Clarify last usage message number in jprint.c
The last usage string is for the end (how surprising :-) ) of the usage
message but if any additional strings have to be added (or removed) the
last number should be changed. Likely it'll be possible to remove one as
the string will be reduced in size once more is done with the tool (and
thus I don't need to see the usage as much) but this will be found out
later.