Update jprint header comments
At this time the tool is strictly being developed by me so I removed
Landon from the comments. It's very possible at the end it will have
been worked on by both of us like jparse was and in which case he'll be
added back but I'd like to have it consistent with what is currently
going on.