Parse -t type in jprint
This is currently done with a bitvector and it might change if it's
discovered that a better way exists. But that won't be known until more
is developed. This was not even the first way I thought of. What will
stay the same is the loop going through the comma-separated optarg but
whether the keeping track and checking of types will be done differently
is TBD later.

Other comma-separated lists in the options will also likely be done in a
similar way but how the information is stored is unknown at this time.

Updated CHANGES.md.