Move jprint option parsing code to jprint_util
This refers to the functions - not the getopt() loop. It feels like this
is more of a utility function. Right now it's only one function which
was also renamed to make it a clear that it's for jprint. It's now
called jprint_parse_types_option() instead of parse_types_option().

This change required a minor fix to the Makefile as well but no need for
make depend.