Start parsing of jprint -l, -i, -N
Add missing -l option itself to getopt().

Add two structs, one contained in the other, for number ranges. This is
a work in progress and might change in details.

Add beginning to parsing of these options. Right now it zeroes out the
structs and returns true as long as optarg != NULL and *optarg != '\0'.

More will be done in a future commit.