Finish parsing jprint options (-b, -I)
Since '\t' is 8 spaces and since the other mode for the options is
number of spaces if optarg is "t" or "tab" set the variable to 8.
Otherwise attempt to parse as a number. If it fails to parse it's an
error.

This completes at least for now the parsing of all options if I'm not
very much mistaken. As long as I did not miss any in the long list this
is done. Once I start adding functionality (the purpose of the tool
itself) some of the options might very well change.