Clarify JPRINT_TYPE_ANY type in jprint_match_any()
It might seem like it should check if JPRINT_TYPE_ANY is set in types
rather than if the types is equal but because JPRINT_TYPE_ANY is a
bitwise OR Of all the types it would be a false positive to check for
say JPRINT_TYPE_INT & JPRINT_TYPE_ANY will be 1 as 1 & 511 is 1. If we
didn't check for equality the test would say that other bits were set
even if they're not (since we will have to check if JPRINT_TYPE_ANY is
set).