Update jprint version for all options parsed
I forgot to do this yesterday. CHANGES.md updated.

I plan to finish the rest (at least known at this time that I can think
of) of the test functions (for the options set) in the coming commits
which will probably be reason enough to update the version again but I
wanted to make sure that this version is specific to all options have
been parsed.