New jprint version - initial test code added
Add option -K which runs test functions and exits 0 if successful and 1
otherwise. This currently conflicts with the exit codes but this will be
fixed in a future commit.

Number range functions added and everything runs fine.

Next is the bit options but this will take a bit more time and
must be done later as I have to head off for a bit. I hope to do it
today also but this is uncertain.