Fix usage strings in jprint
By accident I had put the -K option description in the middle of the -m
and note about that option.

Also it doesn't test the options parsed but rather constraints whether
they were passed prior to the -K or not (even if any were passed it
won't use those) so the description was fixed.