Add code to test jprint -p option
New jprint version - 0.0.9 2023-06-08 - though perhaps this is not
necessary. The next step is to test the -t option bits but this will
happen after a break if it happens today (it should as it's very similar
to the -p option and even uses the same function due to a function
pointer).

The -p test runs successfully.

Also I fixed dead code in the function to check number ranges so that
now both paths are run.