Fix make seqcexit
Some files were missing from the Makefile source and header files
variables so seqcexit was not processing them. The make seqcexit rule
was not run as there are currently other changes in two of the files
that should be in another commit (with different summary) but which can
have in the log that make seqcexit was run but the changes in the file
do indeed have some exit codes changed about.