New jprint version 0.0.10 2023-06-09
Fix handling of number range when the maximum is a negative number. This
fix involved the number range check functions to take a new parameter
which is the number of total matches.

There is one debug log that is problematic with this change but this can
be addressed at a later time, possibly by making more than one debug log
message depending on the range.

Sequenced exit codes.

Fix too long lines detected by make picky (previously undetected due to
the files missing in the variables in the Makefile fixed in commit
3ffe7cf6bbcaa1ee727668f81ab6cd15fe4b9b4e which fixed not only make
seqcexit as the log said but also make picky).

Updated CHANGES.md with new mkiocccentry release too as this seems like
a good place to do so as this is a very important change to the jprint
tool.