Clarify jprint -b option in -h string
The description was a bit unclear. It should say that it prints either
the number of spaces or a tab (and see below on this) between the name
and value.

Previously I had it so that -b t was an alias for -b 8 as '\t' tends to
be (if it's by the standard I don't know) 8 but technically 8 spaces and
a tab are different characters (which can be seen in vim and possibly
coloured grep and other tools in very interesting ways!). This was done
that way to make it simpler but it's not ideal so special handling will
have to be added in a future commit to deal with this. The details are
TBD later.