Fix subtle boolean issue with bit test functions
An interesting problem showed up that prevented the bit test functions
from working right which also prevented the possibility of testing the
-t option. The problem is that the check functions returned the bitwise
AND of the bits set and the correct bit but the bool type is not
strictly true or false so in the test function comparing (as equality)
the expected true or false to the return value of the check functions
returned the wrong answer as (for the example that triggered this
problem it was 2) those functions did not return only 0 or 1 but rather
the bitwise AND!

It was done this way as I'm not used to using bools at all but rather
just checking non-zero so in my head I was not truly thinking of boolean
the way it should be thought about. Thus the check functions now return
the bitwise AND != 0 which means they now truly :-) return 0 or 1.

Add initial testing of -t type. Currently only one test is there. More
will follow another time possibly not until tomorrow or the next day.