New jprint version 0.0.12 2023-06-09 - -I improved
Change semantics of jprint -I to match that of -b and -L as this also is
about tabs or spaces. Perhaps not worthy of another version change but
since I changed it for the other semantics changes of options I decided
to go ahead and do it anyway.

This also fixed the problem of the previous incorrect idea of tab
indicating 8 spaces (used to simplify the code but not strictly
correct).