Partially fix jprint -j option
Now implies -p both.

Bug fix the idea that it requires an option argument. That is the -p
option which already did check for args.

Removed duplicate -j in usage string where the erroneous one was '-j
lvl'.

As -j implies -p both I've moved the parsing of -p to a separate
function in jprint_util.c.

The -j option still has some options it is supposed to imply.