jparse -j now implies -p b -b 1 -e -Q -I 4 -t any
Note that the spec used to say -c but I believe that to be a mistake as
the -j option prints using JSON syntax and -c only prints the count of
matches found like grep -c.

For this update a new function 'jprint_parse_st_indent_option()' was
added to parse -I. I might add that the earlier one for -b is in
function 'jprint_parse_st_tokens_option()'.