Fix jparse locations - resolve issue #595
New jparse version:

The problem was that the column in invalid tokens appeared to be always
1 even when it should have been something else.

For instance given the following input:

    $ ./jparse -
    {"test":false,
    "foo":"bar"},

one would get:

    syntax error node type JTYPE_OBJECT in file - at line 2 column 1: <,>
    ERROR[1]: ./jparse: invalid JSON

Following the questionable spec that one cannot end the last element
with a comma, the ',' on line 2 which is at column 13 is in error since
it's the last element of the file. As can be seen it showed column 1
line 2 so the line was correct but the line was not. Now it shows the
correct line and column:

    $ ./jparse -
    {"test":false,
    "foo":"bar"},
    syntax error node type JTYPE_OBJECT in file - at line 2 column 13: <,>
    ERROR[1]: ./jparse: invalid JSON

Modifying the file a bit so the input is:

    $ ./jparse -
    {
    "test":false,
    "foo":"bar",
    }

one will get:

    syntax error in file - at line 4 column 1: <}>
    Warning: ./jparse: JSON parse tree is NULL
    ERROR[1]: ./jparse: invalid JSON

A question one might have is why is it line 4 column 1 when the invalid
character is actually the final ','? It's because there is more input
available and after the ',' on the line:

    "foo":"bar"

it does not expect a '}' as that is an error in syntax.