Make more explicit usage of no name_arg in jprint
Although no name_arg specified always resulted in an exit code of 0 (if
valid JSON) it was not explicit in the code that it will print the
entire file. There is now a new boolean (admittedly redundant but it
makes it clearer) that is set to true if there is no name_arg. The
boolean is called 'print_entire_file'. This boolean has been added
explicitly to checks even though this is redundant again just to make it
clearer in the code.