Make struct jprint options a struct jprint *jprint
New version of jprint - 0.0.15 2023-06-12.

The struct jprint will hold other information besides the options
including additional structs.

In this commit the msg("valid JSON"); call has been made a dbg() call to
match the formatting of the other dbg() calls.

A note on the level DBG_NONE just to be clear. None of these dbg() calls
will keep the DBG_NONE level. It's only this way for now as I don't want
to have to always specify the -v num option every time I am testing
something but I do want to always have this information during
development. To be entirely clear on this point: all messages currently
are subject to change or being deleted. Messages added later will also
be subject to change or being deleted and this holds until the tool is
more fully developed where it would be a problem to always have the
output. It'll probably start to change once matches are shown or
possibly even the entire file (no name_arg).