Fix check of file arg '-' for stdin in jprint
The check was

    if (*argv[0] == '-') { /* ... */ }

but this is incorrect because image if someone does:

    ./jprint -- -555

It would try and read from stdin when the person either made a typo or
they have for some odd reason a file called '-555'.

Now the entire argv[0] is checked via strcmp(). Why argv[0]? Because
after options are parsed argv and argc are shifted.

Interestingly this was discovered as I was working on a (for now much
like jparse_test.sh) test script with the funny error in the log:

     ... in file -555 ...

which is obviously very confusing and bogus. The test script has another
issue to be worked out but this is more important to get in as the tool
is not yet complete anyway.