Add initial jprint_test.sh script
It is not much different from jparse_test.sh except that it does not
test the strings file (which has a json 'document' on each line). This
is because the tool does not allow a file arg to be a string from a
file.

Something it also does do is run jprint -K test mode.

Whether string support should be added to the tool or whether the script
should be modified so that the strings are faked as stdin are questions
that can be answered at another date.