Add patterns linked list to jprint for searching
Functions to add to the list and to free the list added. Patterns are
appended to the list so that the first pattern/regexp specified is the
first that will be looked for. Whether this will matter in the end is
unknown at this time at least by me.

If a pattern is NULL (should never happen) the function (see below) will
not return. If the pattern is empty an empty pattern is added. This
might or might not be something that should change but typically one
would check that the pattern is not empty before calling the function.
Won't add the same pattern more than once. This check is done in a
case-sensitive manner but maybe the -i option should have a play in this
decision.

When iterating through the args on the command line it will add each to
the list via add_jprint_pattern(), specifying if it's a regexp or not
(meaning if -g was specified or not which will be important later on).
After this it will iterate through the list and report what it's
searching for (later this might be done via a higher debug level or
removed entirely but for now it's done always via DBG_NONE).

The free_jprint() function calls the free_jprint_patterns_list()
function to free everything.