Fix testing of jparse error locations
A number of problems occurred that did not show up at first that
resulted in having to make some changes in the way the testing is done.
In jparse/ the test rule will pass the new option -L to jparse_test.sh.
Only when that option is specified will the script run the error
location tests. This is because the test_ioccc/ioccc_test.sh cannot run
the tests properly nor does it need to run these tests.

Also the way the set up of the error files has to be done is slightly
modified too. This is what has to be done.

First one must generate the error files from the directory
jparse/test_jparse like so:

    for i in test_JSON/./bad_loc/*.json; do ../jparse -- "$i" 2>"$i.err" ; done

Once this is done you must test it. You should be able to run make test
from the top level directory or from either the jparse or
jparse/test_jparse subdirectories. If everything goes okay you can then
add the new files including the error files like:

    git add ./jparse/test_jparse/test_JSON/bad_loc/*.json
    git add ./jparse/test_jparse/test_JSON/bad_loc/*.err

Then commit, push and make a pull request.

Removed a possibly problematic test file and added another one. Error
file actually was added this time.

Updated man page with the new option and fixed details of creating the
right error files. The difference has to do with the subdirectory option
of the jparse_test.sh script and the fact the error files have to be
exact matches.