Change jprint -S -> -M, add -s and -S
The old -S for substring matching is now -M.

Add -S for tool args and -s for tool path. Right now -M is processed the
way -S used to be and -s and -S set a path and args string respectively.
If -S is used -s must also be used because without -s path the args are
useless. Whether this will be done differently is TBD later.