Add flexibility to jprint regexps/regular patterns
With -G regexp one can explicitly enable a search for a regexp pattern
without conflicting with the normal patterns (name_args) or -s. But how
does this work so it's not in conflict?

The patterns struct has booleans to indicate what was specified so that
when -G is specified that specific pattern will be added as a regexp but
after the command line if -g was not specified it will be a regular
pattern. If -s is specified -g is still an error. -G can be used with -Y
but the name_arg will be instead the regexp explicitly specified. This
means that the following two commands would result in the same
functionality with respect to -Y:

    jprint .. -Y int -G foo file.json
    jprint .. -Y int -g file.json foo

Without either -G or -g the -Y would search value without regexps
enabled. If however -G foo and file.json foo is used or more than one -G
foo is used or file.json foo foo is used it is an error. This requires a
special additional check in main() to verify this.

In making the above enhancements I discovered and fixed a bug that
caused the file.json arg to be the wrong argv element depending on
options specified.

This feature allows more flexibility in looking for patterns.