Add mode arg to pipe_open()
As jprint will need write mode for popen() but the other tools need read
mode we have to have a new variable to pipe_open(), a char const *mode,
which is either "r" or "w". A question of why is there not a read+write
mode is that unfortunately in macOS that's "r+" but in linux it's not
supported: it must be either "r" or "w" (some implementations support
additionally "e" for for close-on-exec).

Thus we have to have only one of the two modes. As my understanding is
that for jprint we need write mode, not read or both, this hopefully
will be okay. If we need a version that has both then something else
will have to be done. However as I'm unclear on what the command for
jprint will have to do I am going to either have two streams or else
just leave it as read for now, as it will allow testing of the output of
the tools: it can be changed to write when necessary. This might or
might not happen today.

The pipe_open() diff is:

     FILE *
    -pipe_open(char const *name, bool abort_on_error, char const *format, ...)
    +pipe_open(char const *name, char const *mode, bool abort_on_error, char const *format, ...)