Make pipe_open() flush both stdin and stdout
Based on the information in GitHub (it's not on GitHub and even 'in' is
not really the best description :-) ) it appears that if write mode is
enabled it would have to flush both stdin and stdout. However as
discussed in commit 1353ba61065735eefaaace2f8797e21f55fd42ce it cannot
have both, seemingly, in linux. GitHub comment suggests it might better
be a boolean but it's unclear for certain if it's meant to be both read
and write for jprint. I believed it was both but if so we will have a
possible problem as again it appears that linux does not support both
(unless it's an unclear description in the man page); but even if it
does support both in linux it would likely be "rw" and in macOS it's
"r+" as the string and thus we will likely have a problem anyway.