Complete jprint when no pattern specified or -W
When no pattern is specified or the new option -W (for 'write') is
specified to jprint the entire file is printed if valid JSON. If it's
not valid JSON it's an error and the file will not be printed. In the
case it is printed the exit code will be 0.

This change is important because it is one of the key features of the
tool and it's now complete with a new option to do it explicitly (though
it would technically conflict with a pattern specified so if a pattern
is specified the -W option is ignored and maybe it should be a command
line error?)!

New version is 0.0.19 2023-06-16.