Change jprint -W to -o for output file
This seems like a better letter to use. The issue of whether or not it
should be a command line error if -o and any patterns are specified has
not been addressed but my suspicion is it should be an error and
probably will be made into one. This is TBD at a later point in time,
perhaps rather soon.