jprint now executes -S path if executable
Some important points to note. First it runs it via shell_cmd() to check
the exit code. This ends up printing out the output but since the exit
code was noted as significant it will do for now.

If it's non zero it will be an error. Otherwise it opens a write-only
pipe. If the pipe is not NULL it will be processed at a later time once
I am clear what needs to happen. If the pipe is NULL it is an error.
This means that the output is printed twice.

Second point of interest is that the -A args option I am not entirely
clear on whether they should be options that are ended by '--' or if
they should be args to the program. It might not matter even but I'm not
sure. Nevertheless right now it's used as options which. Even so one can
do for instance:

    $ cat foo.json
    {}

Specifying no pattern to search:

    $ ./jprint -S /usr/bin/grep -A '{' foo.json
    debug[0]: set tool path to: </usr/bin/grep>
    debug[0]: set tool args to: <{>
    debug[0]: maximum depth to traverse: 256 (default)
    {}
    debug[0]: no pattern requested or -o, will print entire file
    {}
    {}

..which prints the full file (because of no pattern specified) but runs
grep searching for '{' in the file too. If one specifies a pattern it'll
look like:

    $ ./jprint -S /usr/bin/grep -A '{' foo.json '{'
    debug[0]: set tool path to: </usr/bin/grep>
    debug[0]: set tool args to: <{>
    debug[0]: maximum depth to traverse: 256 (default)
    debug[0]: name pattern requested: {
    debug[0]: adding name pattern '{' to patterns list
    {}
    debug[0]: searching for name pattern '{' (substrings ignored)
    {}

..printing the last part only once as the tool does not search yet and
no full file printing was requested.

One could replace grep with any other tool as well even a shell script
that exits 1.